"11:15am. I slept for a while today. Let me think a bit.
I might have given up too fast yesterday. There are some things that disturb me about how the whole thing went. There is no doubt that just slapping the linear AC on top of the action predictor features does not work. But look at it from another angle - if that does not work why would I expect that GAN trained predictor would work either?

Maybe the correct choice would have been to combine full training of the actor and the action predictor. One step for the first and one stop for the other. What about the value net? Who knows.

It is clear to me that there are clashes in the methods. I thought that things would smoothly work together, but I was completely wrong. For combining the action predictor features, it is clear that things have to be made to work together and that is not good.

11:25am. https://openreview.net/forum?id=cvNYovr16SB
Unsupervised Active Pre-Training for Reinforcement Learning

Abbeel is the second author of this. I'll read this paper.

12pm. Damn it, this is so difficult. Yes, I know the duality gap GAN method, but all this is so complicated. I managed to master tabular CFR and extended to NNs, but I do not have a coherent perspective on how it should all be brought together.

I should learn the lesson from this. If I want to do research, I should focus on that. Maybe if I did that for months and years I'd managed to figure out a coherent framework for combining GAN training with RL.

12:05pm. No, it is not enough. The paper I've linked to is complex, but my own stuff would not be much better.

I have a intuition that prediction training would help things because the brain does it, but I do not have intuition for how to best compose it with RL. I want to do things because the brain does it without actually knowing or understanding how or why the brain actually does things.

12:15pm. In the story of the wizard I am thinking of, he started off with spells to intimiately read the activity of souls before moving on to modifications.

I can struggle and persevere, but could I have come to the present level on my own? By that I mean probably just the current tabular CFR algorithm. Probably not.

And I absolutely can't come anywhere to contructing agents with even animal level intelligence on my present level.

More knowledge would help. More understanding would help. More computation would help. More intelligence would help.

I can't break to 4/5 in ML just yet. What I have now is the quintessential 3/5 - I know how to do naive RL with NNs and not much more beyond that. I do not know how to augment it with unsupervised learning and deep exploration. Not that there aren't hints pointing in that direction, but that is a far cry from having a coherent mental framewrok of how things shuold work.

12:30pm. I am fighting on too many fronts. Maybe the hints I have are enough to speculate, but they aren't enough to dive in.

Instead now that I've come to this point how about I firmly set my foot down and just fight economically?

I can see that once I put in the transformers, trying to make unsupervised GAN training work with it would be hugely time consuming.

It is not just a matter of programming, it is clear enough how long it takes me to run the experiments. To get good data here, Leduc would not be enough, I'd have to run it on Holdem for a lot longer.

12:35pm. Algorithms are derived from the understanding of structure. I do not understand the problem of optimizing high dimensional objects so it is no wonder I can only do this much.

I am sure that the actual algorithm that bridges all the gaps will be profound. I do not feel it will be particularly complex. It will certainly be a lot simpler than piling hack after hack in hopes of getting things to work.

12:40pm. I am not that desperate for it just yet. It is fine if I could reach the appex of the present mountain.

For all I know, maybe just transformers would be enough to crack Holdem even without a need for a curriculum. And if not I'll just do the curriculum.

12:45pm. I need power. 3/5 should be enough to crack any of these simple betting games. I should not lose sight of my present goal.

Crack these games, make money off them, get the neurochips. That will allow me to train on 1,000x more rounds in the same amount of time that I can do now. I should do that, and completely raze the online gambling dens to the ground using my superhuman gambling machines.

Maybe this will take me all of my time, and I won't have time to play around with GANs and research new algorithms. If so that is fine. In the timeframe I have set out, there is not much time for me to do any great research myself.

12:50pm. Succeeding at what I've set out to do here would be enough of an accomplishment on its own. Eventually, the neuroscience guys will stop being useless and figure out what the brain itself is doing. No way can I do that with the sheer power of my imagination.

I have no doubt that much better, lower variance ways of doing what I am trying to do now will be found out.

The depths of skill and capability that I have now cannot be compared to what will be possible in the post-Singularity era.

12:55pm. I am just a human, the power of my imagination is not enough to make the world turn where I want it.

1pm. I do not have a great wellsprint of wisdom to draw out the relevant algorithms.

It would be for the best that I focus on stealing them instead. Right now I am just competing against the world and nature instead of being smart about it. If I had a billion dollars and want to figure things out at all cost, I'd put that money into brain recording.

1:05pm. At any given point in time, there should be ways of making money of existing ML and programming technology and I should just focus on that. Today board and betting games are what is viable. Tomorrow, it will be games like Dota and Starcraft. I need to live in today."