"10:30am. Let me make a show of appreciation here and plug a story I've been reading recently. It is why I've been going to sleep at 2am for the past few days and getting up late.
https://www.royalroad.com/fiction/40150/pith

Most of the GB stories out there are garbage and I should know since it is my hobby to read them. I've enjoyed Magic Smithing and the Hive Queen on Royal Road, but Pith is in a league on its own from characterization to world building. It starts off with an interesting premise - the MC tries stealing a body to transfer her soul into, and keeps being strong throughout.

It is currently on hiatus because the author has developed repetitive strain injuries due to too much writing, but he otherwise is eager to continue it. 5/5.

10:45am. Let me chill for a while, I do not feel like rushing it. There are too few hours in a day for me.

Today's agenda is to get the Python scripts adjusted and run a transformer for the first time.

11:25am. Almost done with chilling. Let me just read a bit more and then I will finally start. Though, maybe I should just have breakfast instead of starting and droping it.

https://www.reddit.com/r/MachineLearning/comments/oif5dp/d_adding_guassian_noise_to_discriminator_layers/

Let me also plug this thread. I am shocked that nobody tried this before.

11:30am. I am want the deal to sweeten a bit before I try unsupervised learning. Trying upside down RL, soured me on the prospects of composing different methods.

What I have in mind is not enough. I want something new, some better principles for how to compose disparate tasks and methods.

I could resume Simulacrum tomorrow, but just having the story ideas is not enough for me. I want the algorithmic advancements to inspire me, and deep learning is not enough. The way nature does mind engineering is way beyond anything I could do with backprop. I want to understand it, but it is beyond my capacity to infer. So I'll dig in at my present level and wait for the opportunity to come. I want to look at the algorithms nature uses and take from that the nature of reality. Necromutation might be transcendence, but that view is too low resolution. I need more to motivate me.

Let me do some programming.

Ah, right. The next things are `create_args` modules.

```
inl main () =
    open leduc
    inl vs_self = train.vs_self leduc.game
    inl vs_one = train.vs_one leduc.game
    inl neural = train.neural agent.neural_leduc.extractor agent.neural_leduc.schema()
    inl uniform_player : ra u64 (pl2 card action * leduc_state * u8 * a u64 action) -> a u64 (log_prob * action) * (a u64 r2 -> a u64 r2) =
        agent.uniform.policy
    inl tabular =
        inl create_policy : agent.tabular.agent_dict (list (choice2 card action)) card * bool * bool * f32 -> ra u64 (pl2 card action * leduc_state * u8 * a u64 action) -> a u64 (log_prob * action) * (a u64 r2 -> a u64 r2) =
            agent.tabular.policy agent.neural_leduc.extractor'
        inl create_agent () : agent.tabular.agent_dict (list (choice2 card action)) card = dictm.empty
        inl optimize : agent.tabular.agent_dict (list (choice2 card action)) card -> () = agent.tabular.optimize
        inl average : agent.tabular.agent_dict (list (choice2 card action)) card * agent.tabular.agent_dict (list (choice2 card action)) card -> () = agent.tabular.average
        inl head_multiply_ : agent.tabular.agent_dict (list (choice2 card action)) card * _ -> _ = agent.tabular.head_multiply_
        namedtuple "Tabular" {create_policy create_agent head_multiply_ optimize average}

    record {vs_self vs_one neural uniform_player tabular}
```

Here is the `create_args_leduc`. Now let me do the other one.

```
inl main () =
    open hu_holdem
    inl vs_self x = train.vs_self (game x)
    inl vs_one x = train.vs_one (game x)
    inl uniform_player : ra u64 (pl2 i8 action * ({hand : i8 * i8; id : u8; pot : i16} * {hand : i8 * i8; id : u8; pot : i16} * a i16 i8 * i16 * bool) * u8 * a i16 action) -> a u64 (log_prob * action) * (a u64 r2 -> a u64 r2) = agent.uniform.policy
    inl callbot_player : ra u64 (pl2 i8 action * ({hand : i8 * i8; id : u8; pot : i16} * {hand : i8 * i8; id : u8; pot : i16} * a i16 i8 * i16 * bool) * u8 * a i16 action) -> a u64 (log_prob * action) * (a u64 r2 -> a u64 r2) = agent.holdem_callbot.policy
    inl neural stack_size = train.neural agent.neural_holdem.extractor (agent.neural_holdem.schema stack_size)
    record {ui train = record {vs_self vs_one uniform_player callbot_player neural}}
```

Yes, this is fine. No type errors in the workspace. Let me try compiling it. ...It went through. No type errors during peval.

12pm. Focus me, stop taking peeks at the Pathfinder thread.

Now that I am done with the Spiral side...

12:25pm. I know I should be heading straight into Python scripts, but I am having some inspiration.

It might be possible to get rid of the actor by instead taking the max over the values in the critic. The way to derive a policy would be to pass an noise vector along with the inputs. Varying that noise vector would give an implicit policy.

Why would that might be worth trying out? The reason is that all the variance of importance sampling that CFR does would simply go away if the sampling and the policy distributions were one-hot.

Right now I understand very well that the algorithm is unsuitable for anything but short sequences because of that.

12:35pm. Let me stop here."