"6:25pm. I've been reflecting all this while. Without a doubt, I'll finish the player, but I am getting ahead of myself. Money is an entirely separate matter from that.
Even if I had the elite player, my odds do not look good. Can I automate the player so it uses the Stars or any other client on its own. I doubt it. Just how much effort would that take?

Furthermore what is more important, a billion dollars or a 5/5 skill level in ML? I'd pick the later in a heart beat.

I've lost track of what is important. I've lost track that what is really important is the labor I've put into this, not the eventual success. I've gotten arrogant from my success and started thinking that I've made it when I haven't even begun.

6:35pm. Ah, I want to win even a little so as to fulfill a long unmet need in me. But I cannot forget that this is not a need that needs filling in the first place.

Making these agents is important because that is the only way to improve my ML skills. But making money is a different thing. Maybe it will come, maybe it won't. Forget it. Forget that part of the goal. If I want money that badly I should get a job. What really matters is getting better, accumulating skill and knowledge through continuous experimentation.

I've gotten completely deluded, thinking that somehow getting this far means I can take a break. If I want a break I should just take it.

I've gotten weak thinking I could get the benefits of ML without the effort.

6:40pm. Money is a spook. Forget making it next month or even year. I should just take my sweet time practicing ML and it is going to come rolling in on its own eventually.

If the past is making you insane in a bad way, ditch it. If the future is making you insane in a bad way, ditch it as well. And both of those are definitely pushing me in a bad direction. My regrets and longing will give me nothing in the end.

Only the small goals I aim towards every day will have any sort of benefit.

6:45pm. By no means am I close to ready to use the agents in the real world. In no way do I have the skills for that. I am just close to getting them to train for the first time, and have a good exploration method to boot. Once I do that I'll have enough at most to write an ML paper about my methods.

7:05pm. It would actually be super difficult to do UI automation by training a neural net. Things like clicking buttons, reading the pot sizes, bet sizes directly from the screen. It seems simple, but would be both beyond my computational budget and ability to implement.

Most likely I will have to implement some static scheme that relies on careful positioning of things. Or play myself with the NN as the advisor. But those kinds of activities are exactly the sort of thing I am not interested in doing. If I had to do that, a job would be a better choice.

So then making an app and selling it might be a quick way to net me some earnings. If that is my plan then I could realistically count on making some money.

Alternative, hack the client itself somehow and get the sizes and make the commands based on its internal states. But the clients frequently update. And they will no doubt have protection against that sort of thing. All I have to do is get caught once and then I am out. Big money makers would use proxies, but I do not have access to that.

Several years ago when I was doing research, I found a community of poker botters using scripted bots. They had an UI interface to the sites. But I read it was brittle and could be thrown off by one pixel difference and so on. I'd be fighting the dens if I went down this route.

7:25pm. I am going to have to weight my greed and ability carefully against each other. I am also quite psychologically fatigued from the neverending struggle.

I am not a 500 year old reincarnated demon with a heart of steel.

It is not like I have a post-Singularity perspective on the actual learning algorithms. Right now I have the first realistic thing that could possibly work.

7:35pm. If I am really serious about curtailing my arrogance, I should swallow my humiliation and accept my limitations for the time being.

7:50pm. Really, selling my agents is the middle ground between getting a job and outright using them. It is the only compromise acceptable that does not set me on a different path.

8pm. Sigh, I am just too weak right now. Let me aim for that. This will be my working plan as I develop my skills. Of course I'll check out the prospects for UI automation, but won't expect much from it.

8:10pm. It has been 6.5 years since I last upgraded my rig. Getting that newest RTX 3090 would make a difference in my capability to train larger models. I should do what is needed. Selling the initial versions of my agents should allow me to fund the development of better ones.

Still, just where will I get the algorithmic improvements past this point? I have no idea. Maybe I'll be able to branch out to different versions of the game. If I start selling, I'll have to compete with other people in this bussiness. Who knows how good my competition will be.

8:25pm. I should be determined to do this. I will look into using the agents directly, but if that turns out to be too hard I'll go the entrepreneur route. People who sell weapons make more money than those who use them anyway. I'll have to play my cards right, but I should be decisive in my desire to do either. I cannot falter here."