"6:35pm. Actually, I really should just let the thing train now and over the night. Let me go for it. That way when I wake up I'll be able to tell right away whether my training works. If I close here I'll just be wasting time. I might as well open the case in order to air it out.
7:15pm. Tried playing vs the checkpoint 4 player. Back to being an aggrodonk.

I have no idea, I want to pivot, but let me instead keep running it. I do get the feeling that I am not doing it right, but I can't overcome the lack of hardware so easily. If I get no results by tomorrow, I'll change the way the training is done.

7:35pm. Let me try agent 5.

7:40pm. It is still quite bad, but at least it looks like it is learning something. It is not a dumb aggrodonk. Instead it makes small bets, and sometimes calls down with pairs. It is still highly exploitable as it will keep reraising and then folding 90% of its stack again and again.

But this behavior is still much more refined than folding all the time or making massive bets and calling down with anything all the time.

It is not out of the question that it might turn into something good eventually. It is worth seeing whether it improves in the coming iterations or if it degenerates back into an aggrodonk. If that happens, I'll have to retry the training with bigger replaement gaps. I am afraid that if I make the gaps too big the action prob softmax will saturate.

Though I've changed Hellinger distance to log softmax, I am still not sure if normed square or softmax are better when feeding the values directly. I can't assume either thing.

The training here is so unstable that I have no choice to just pick one and stick with it.

Running sweeps is way out of the question. This is the curse of having small compute resources.

8:30pm. Let me try agent 6. It is not an aggrodonk, but it barely moved from the previous one. It keeps folding to reraises preflop even with very strong hands.

I poker really such a difficult game that this is the best they can do?

Actually their behavior is pretty interesting. Why would they put so much money in only to fold at the end?

This reminds me of the behavior of the Poker Academy Pro 2 bots. They too were very liberal with raises, but would fold unless they had top pair. It made them very exploitable. To think I would see this behavior again.

8:45pm. Well, maybe the bots will get to a decent level over the night.

It is too bad I lost the game somewhere, I remember looking for it and couldn't find a download for it either. One interesting part is that it had that GTO limit HU player. That one was tough as nails, I did my best, but could not find a way to beat it.

9pm. The way the agent 6 acted, it would push strong hands for longer before folding. This is quite interesting, and it is not like agent 5 which would push everything before folding. Agent 6 would only push weak hands to about 30% of its stack.

So it is learning. I probably won't have to increase the replacement length.

9:25pm. It is weird. It refuses to call down even with top pair for some reason. Trips are fine though.

9:30pm. Oh 7 is done. Let me see how it progressed. I am glad I let it run, this is fun.

I has similar behavior to before, but it has learned to call down with KK preflop.

Folded AA on the flop. It is dumb as a rock.

10:20pm. 5 iterations till 8 is baked. What a difference a even 100x fold in performance would make. It would be at 800 by now.

The sheer amount of time it is taking it to get to a halfway competent level is insane.

I could theory train this for months, but do I feel like having my GPU burn away for 24/7 during that time. I also like to turn my PC off during the night.

I'll do that. Once it hits 8 I will abort. Absolutely nothing in the world is worth distrupting your sleep for.

10:30pm. Let me give 8 a try. Hmmm, did I forgot to try out 7?

10:45pm. It really is getting better. Right now it actually knows to call down with pairs. Apart for the sake of showdown, it never seems to call to see the next street, it only ever raises and folds. And it does so in a very exploitable way, but gradually it is getting better.

I want to see how good it can get. Let me leave it overnight after all."