"8:15am. I got up early today. This is good as it will be more comfortable doing the chores before the sun gets up. Let me do it right now. Later I'll work on adding token literal suffixes to the language.
10:10am. So there are about 6.5 rows that need filling and at the rate of 0.5 a row a day I'll be done with the winter preparation is 13 days. Good.

I've had time to think about what I want to do today. First of all, this TensTorrent opportunity is an ideal test of Spiral's perceived worth. I know well enough myself that if the company adopted the language it would be a drastic improvement of the ease of use of their device. Even the old Spiral was like that with Cuda. The newer v2 would be a heaven sent opportunity for them.

But it is not like this encounter is easy for others to recognize.

I've decided that I am ready to start the job hunt. I do not want to depend on others. I sent the email to the CEO yesterday and it has been enough time for him to look at it and skip it over.

I'll get the job and make money with my own hands with which to buy the device.

But in return, if it turns out that Spiral cannot give me these minor advantages of getting chips as a reward for my good work of the past 6.5 years, then what am I keeping this public show of will for? There is no need to be a spectacle.

If I were to do a TensTorrent backend, other people could greatly benefit from it, but I myself would gain nothing from other people benefiting. I do not want the company to free ride on my work. And since I am doing a poker agent, and since the chip I'll surely get it to train, doing that in public would just depress my own earnings.

10:20am. In other words, it is like this - if Spiral is to be open source, other people will have to contribute more significantly. 3 people donating a total of 120E over several months is not enough of a benefit.

Right now, my RL research is in a lurch, and if I have to do everything myself, I might as well save myself the effort and take in further Spiral development private.

10:25am. I'll make it a race, if I get a positive response from the CEO and he grants me a Grayskull card, I'll do the backend and continue the project I am doing in public just as I have up to now. If not, that is proof enough that there is no interest in Spiral from AI companies and I should spend my energy more wisely instead of trying to catch people's attention.

Ultimately, there is a purpose to using these commits as a diary, and doing the PL sub reviews. If it is not being fulfilled and the show of will is pointless, then I should recognize that reality and embrace the shadows.

10:35am. Turing has been spamming for remote jobs. Let's see what they have. I'll spent today applying to a few place.

10:50am. The site is already pissing me off, it says it accepts .doc files, but then rejects them once they are uploaded. I guess I'll go with a PDF.

11:25am. The site is interesting. I fill out the resume, it uses ML to dig out keywords then I proceed to fill out the relevant work experience, and now it has a bunch of tests waiting for me. For some reason it is really pushing SQL on me.

Well, whatever. I can pick that up in a day.

One thing I do not want to fuck up is the C# test. I am actually an F# expert, not a C# one. I think I should play around with C# for a few hours before taking the test because I do not want to be fooling around with its syntax in a timed situation. When I am doing things at my own pace, dealing with C# is no problem, but I should take a refresher.

The first thing I have to deal with is tell them about my work experience.

11:35am. There is gaming potential for this site. If I wanted to be 100% sure of passing the test, I could make a fake persona and check them out ahead of time.

Even without looking at them, I am almost sure to pass what I have experience in. They are unlikely to make them too hard. But...I am tempted to bend the odds even further in my favor.

11:40am. No I am overthinking it. They are likely to be super easy.

Let me have breakfast here, then I'll do the work experience assessment. After that I'll play with C# for an hour or two to warm myself up and take the test. If I can pass that, I am set. If I fuck up for some reason I'll just look for work elsewhere. If I really want to push Turing I could just reapply using my father's id. That would make things complicated though.

It is not really that important either way. If there is one Turing, there should be another elsewhere in this space."