"8:25am. Let me do the chores here.
10:40am. 6 rows left.

10:45am. Time to chill. If I had money I'd buy air conditioning for my room.

Let me be realistic. Whether I take Spiral private or not makes no difference either way since I am the only user. I should just forget those lines of thinking and continue to use the project as my diary.

The 2021 attempt at RL has essentially failed. And once I get a job, I doubt I'll give a damn about poker anymore. The sweet lul of having a steady stream of money will addict me to it, and any excess will go to the stock market. My feel is pretty good, right now thing couldn't be better with QQQ near all time highs, and so it is time to stay out.

ARKK and TAN are not really going up, so it is time to avoid them too and wait for a month or two to see if they show strength. If not avoid even longer.

10:50am. I think the current method I have could train a decent agent, but not in my home GPU. It is actually quite arduous having the GPU whiz around during the day and night.

If I had money I could rent a cloud and do it that way. I should be cultivating my cloud computing skills.

Every one of the AI companies will be selling to the data centers, so even if I can't buy these cards directly, I'll be able to rent them.

10:55am. The main problem is that I am 3/5 in ML and every attempt to break into 4/5 so far has failed. It reminds me of high school and how I've tried to solve those olympiad level problems to no awail.

To go beyond backprop I need to go beyond calculus.

At first there were integers, then came the the floats. Compared to doing discrete optimization, doing it over a differentiable manifold is much easier, but it is not enough. In order to break through to 4/5, a genius on par with Newton is needed.

Just like floats (well reals and rationals) were an evolutionary step over the integers, I think what we need is some kind of novel data structure that will lead to new ways of thinking about learning and computing.

Having backprop never resolved any of the issues with discrete problems. And the next level method is not going to resolve the issues with backprop.

11am. Reversibility? I should learn about reversible computing, but that perspective have too many issues to be viable. Still, right now I am 1/5 in that domain, so it like a elementary schooler making judgements. I can't trust it.

11:05am. I've really pushed deep learning to its limit.

Maybe if I'd have gotten that Grayskull card, I could have continued the quest for a while longer, but I'd run into the same scaling issue on something bigger than HU Holdem. It does not require much for DL to start choking.

I should just retrench. Chasing your goals is a virtue, but enduring hardship is a sin. There is a balance to be struck between the two. No goal that is worthy is without hardship, but seeking hardship for its own sake is foolishness.

Let me update my plan:

* Make paid programming work the main money making activity. Forget poker for the sake of that purpose.
* Once I have income, learn about cloud computing. Use a cloud to train the agent using whatever AI chip is available. Enjoy having my room cool and quiet. Ride the new hardware wave that way.
* Try to find a way to break into the 4/5 ML realm. This will require other people to get inspiration from nature and share it with me. Since I have no better ideas, I should study reversible computing and see if cultivating that gains me anything.

That last is the most important, but breaking realms takes new principles, algorithms and perspectives. Despite all my effort over the past 6.5 years, I have nothing significant to contribute to ML.

11:20am. I need something significant to make that crucial breakthrough. It is not like I myself couldn't have invented calculus. The principles behind it aren't hard to understand.

But if I had just started with the rules of backprop, it would have been really hard to work through to the principles. Nothing could have told me the rules are derived by making an infinitesimally small change to the objective function. I've would be stuck in a tarpit of rules.

11:25am. Since it will take a while to find a job, while I search I should work on Spiral more.

TODO: Just like for modules, show the result of the type hover of every application involving a symbol.
TODO: Comments for prototype functions.
TODO: Interpolated strings.
TODO: Highlight unused vars.
TODO: Autocomplete.
TODO: Make a VS Code theme for Spiral.
TODO: $"!!" and $"``" should be valid syntax in macros.
TODO: Make literal suffices (such as `i64`) be highlighted differently from the rest of the number.

11:35am. Some of these are easier than others. I guess I'll focus on Spiral since I have nothing better to do. Maybe I could look into reversible computing as well.

TODO: Just like for modules, show the result of the type hover of every application involving a symbol.
TODO: Improve the way modules are printed.
TODO: Comments for prototype functions.
TODO: Interpolated strings.
TODO: Make a VS Code theme for Spiral.
TODO: $"!!" and $"``" should be valid syntax in macros.
TODO: Make literal suffices (such as `i64`) be highlighted differently from the rest of the number.

Let me do these in my off time. Highlighting unused vars and autocomplete would require writing passes specific to them, and I do not want to bother with it.

11:45am. Yeah, it is Saturday, so I should rest for a while. Striving forward with all your effort is only viable when you have a direct goal.

I'll take the skills I have acquired over the past 6.5 years as my reward.

11:45am. Let me finally read the Inglis bath scene instead of whirring my thoughts around.

12:30pm. Breakfast time.

https://spectrum.ieee.org/the-future-of-computing-depends-on-making-it-reversible

My first goal is to figure out whether there is something to study here at all."