"6:30am. I am up. To think that I both had time to lounge in bed and do the chores by this point. Let me check out the mail and then I'll put the v2.1 docs through the grammar checker.
No, nothing. Yeah, my working estimate that nothing is going to come from this. There are too many stories of people looking for jobs for half a year. I am not going to make it this way.

6:35am. Let me also push out the new Spiral.

6:40am. Interesting how many errors are in the docs. The Google Docs tool is catching a bunch of them, and it is not from the newer segments either!

7:10am. Fixed a bunch of errors. I am surprised how many they were even in sections that should have been checked previously.

Let me commit here so the readme shows up on the main page. After that I'll chill a bit and practice reciting.

I'll also take the time to study the /r/cscareerquestions sub. I noticed yesterday that there is a resume thread, and it gave me inspiration how to redo my own. I should rewrite it and put my work experience on top.

7:15pm. This I'll leave for later. Today my main priority should be to learn to speak better. I need to level up that skill. Then I will be able to pass that Toptal test. After that will come the actual programming tests that I need to pass in order to be eligible for work.

I'll have to give that my utmost. That AI chip will not land itself in my lap on its own."