"1:35pm. https://www.snagajob.com/blog/post/part-time-hours-vs-full-time-hours > Part-time employees are always paid by the hour. That’s why they are usually responsible for “clocking in” and “clocking out” at the beginning and end of their shifts.
Interesting. I had no idea that this was the case.

1:50pm. Done with breakfast. 40m till the meeting so I am not going to be starting work on the backend right now.

Since I've passed the HR test for Tessian, I have 4 active threads at the moment, so I have no more incentive to keep applying. Maybe I'll get more tests in the future that need taking care of.

2:20pm. It is almost time.

2:25pm. I just can't relax. Ok, let me join the meeting.

3:05pm. That went well. He says he will get back to me tomorrow after checking with his manager.

Ok, this job if I took it purely what he is saying is really an operations research kind of job hidding under a compiler one. The chip they have does not have memory apart from the registers, and does not allow spilling for that reason. All the registers are statically allocated ahead of time and the main problem is to optimize them. Normal compilers also optimize the artithmetic operations, but the chip has a lot of ALUs so that is not a problem there.

https://en.wikipedia.org/wiki/Arithmetic_logic_unit

I had a moment where I forgot what ALUs were (arithmetic...linear unit?) and had to resist the urge to ask what they were.

The big problem for them is that the classical compiler algorithms are not applicable for their kind of exotic hardware.

3:15pm. He mentioned the P4 programming language. I should check it out.

https://en.wikipedia.org/wiki/P4_(programming_language)

This is what I'll be working on if I join their team. More precisely the backend for their network chip which compiles to P4. Thought there maybe a way to angle it towards working more on the front end side of things.

https://p4.org/

I suggested I could do a Spiral backend for this in my usual fashion, but now I am not sure.

3:20pm. Ok, I do not feel like programming. Applying to these companies is like an adventure, so I might as well spend some time admiring the sights. Let me study p4 for a bit.

https://github.com/p4lang/education/blob/master/GettingStarted.md

Let me watch the video tutorials here.

3:55pm. Yeah, I am exhausted. I do not have the focus to really watch the videos. This kind of hardware is far from my usual interest. Not that it would be a bad job to have.

I can't really focus on the videos, but let me just go through them anyway.

4:25pm. https://youtu.be/cvDtVobw9wE?t=369

Ok, I definitely won't be able to make a Spiral backend for P4. The languages are too different.

I am not sure I get this example. Why does this switch the ports?

5:10pm. Done with the break. I am in a daze trying to process the conversation. I watched only two videos, and did not get the programs, but I have a much clearer picture of the language is about.

I admit I was at a loss when he asked me what I could do, but now some ideas are starting to form. If I get the chance to go further I'll elaborate them. Let me watch another video.

5:35pm. I am just pacing around the room in a daze. Now that I've had the time to think about this job, it might be better than I thought at first glace.

6:15pm. Done with lunch. Let me watch the video.

7pm. Only finished the 3rd video just now. Basically, this job is actually on my current path. When I applied, I thought meh, router switches...that has nothing to do with me. But if the job is about graph optimization, that changes things completely.

I mean think about what I want to do - implement the poker game and the algorithms on an AI chip. Instead what if I introduced randomized algorithms to the P4 backend, and then implement the compiler optimizations on the AI chip that would be quite a similar thing.

I'd need to change my specialty from RL to graph optimization, but it is not like I care about current day RL that much. It just seemed like the most viable path, and graph optimization is not something I'd even considered. But I do have useful experience from the two discrete optimization courses that I've taken. I never thought that I'd ever find use for it.

Here are 3 main points that I'll suggest, assuming they want to go for another round of interviews. When I was asked what I could do, I could only mumble something about AI chips.

1) I'll propose a grand project to do the graph optimizations on AI chips. I'll take the randomized algorithms that I've implemented on CPU, port them on chip, massively parallelize them and investigate the use of NNs to guide the search. Using NNs in such a manner would be insane with just CPU+GPU as it would blow up the compile times, but these AI chips will have local memory and specialized operations that will make running them very fast. They'll also be suitable for graph optimization work. If this works and it probably will, the optimization will be very ferocious and none of their competitors will be able to match up. It won't be easy to copy this approach, since there aren't too many people with my skillset in both compilers and ML.

2) I'll open a line of research into randomized algorithms, and try them out the problems. No doubt, they've accumulated a large dataset of programs to test their optimization algorithms on, and I will take a look at giving the metaheuristic approaches a try. Ultimately, classical algorithms are heuristic approaches and only represent the upper bound on whatever measure they are using. I had really good experience replacing constraint solving with metaheuristic approaches during those courses and I'll want to lean on it. Though the viability of this depends on how much their customers would prefer for to wait for the optimizer to do a better job instead of rewriting the program so it does something that uses less registers.

3) Initially, what I will try is rewriting some of their codebase in Spiral. This will both serve to get me acclimated with their compiler and demonstrate what large cuts in the amount of lines of code are achievable using a functional language like Spiral. It will give me an opportunity to get some actual users and start weaning their imperative programmers off C++ and into something better.

I should fight the urge to be bombastic and go from last to first here.

7:25pm. Ultimately trying out optimization algorithms on novel problems is exactly what I've been doing for the last 6.5 years, so this kind of job is a good fit.

Alternatively, they could say no to functional programming and randomized algorithms, but in that case they won't get their money's worth out of me.

I really hope I get the chance to present my case, and that Viktor's manager doesn't just decide to skip. It would be too bad if that happens. I'd gotten a little hyped about this, so I'll feel dissapointment if it slips away.

7:35pm. Hmmm, if I get a reject I'll actually write out my plan and send it to him to see if that changes his mind. Getting my resume tossed into the trash bin after a 5m glance is one thing, but given that I've gotten an face-to-face interview we have some connection. I should not give up on it too easily.

7:40pm. Ugh, let me finally watch that 4th video and I will call it a day.

https://youtu.be/KlEi87XYMBE?t=7
P4 Tutorial 4 - Basic Tunneling Application

It is very rare for me to watch programming videos and not understand what is going on, but has been the case so far. I'll have to try out P4 directly."